Move unit tests to github CI
